pyproject.toml
[tool.poetry]
name = "yify-scraper"
version = "0.1.0"
description = "Yify movie scraper"
authors = ["DuneRaccoon <benjamincsherro@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.28.1"
beautifulsoup4 = "^4.13.3"
libtorrent = "^2.0.11"
leaky-bucket-py = "^0.1.3"
loguru = "^0.7.3"
pydantic = "^2.10.6"
torf = "^4.3.0"
torrentp = "^0.2.3"
pydantic-settings = "^2.8.0"
croniter = "^6.0.0"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
